fix: change endpoint of passkey registration api
