+++ package.json
@@ -37,6 +37,7 @@
     "@nestjs/terminus": "^11.0.0",
     "@prisma/client": "^6.3.0",
     "@simplewebauthn/server": "^13.1.2",
+    "@simplewebauthn/types": "^12.0.0",
     "@simplewebauthn/typescript-types": "^8.3.4",
     "bcryptjs": "^2.4.3",
     "cache-manager": "^6.4.0",
