+++ src/auth/auth.service.ts
@@ -16,17 +16,11 @@ import ms, { StringValue } from 'ms';
 import { AuthRepository } from './auth.repository';
 import { LoginDto } from './dto/req.dto';
 import { LoginResultType } from './types/loginResult.type';
-import { UserRepository } from 'src/user/user.repository';
 import {
   generateAuthenticationOptions,
-  generateRegistrationOptions,
   verifyAuthenticationResponse,
-  verifyRegistrationResponse,
 } from '@simplewebauthn/server';
-import {
-  AuthenticationResponseJSON,
-  RegistrationResponseJSON,
-} from '@simplewebauthn/typescript-types';
+import { AuthenticationResponseJSON } from '@simplewebauthn/types';
 
 @Injectable()
 @Loggable()
@@ -44,7 +38,6 @@ export class AuthService {
     private readonly configService: ConfigService,
     private readonly redisService: RedisService,
     private readonly jwtService: JwtService,
-    private readonly userRepository: UserRepository,
   ) {
     this.accessTokenExpireTime = ms(
       configService.getOrThrow<string>('JWT_EXPIRE') as StringValue,
@@ -127,85 +120,10 @@ export class AuthService {
     return this.authRepository.findUserByUuid(uuid);
   }
 
-  async registerOptions(
-    email: string,
-  ): Promise<PublicKeyCredentialCreationOptionsJSON> {
-    const user = await this.userRepository.findUserByEmail(email);
-
-    const options = await generateRegistrationOptions({
-      rpName: 'idp',
-      rpID: this.passkeyRpId,
-      userID: Buffer.from(user.uuid),
-      userName: user.name,
-      excludeCredentials: user.authenticators.map((auth) => ({
-        id: auth.credentialId.toString(),
-        type: 'public-key',
-      })),
-    });
-
-    await this.redisService.set<string>(user.uuid, options.challenge, {
-      prefix: this.passkeyPrefix,
-      ttl: 10 * 60,
-    });
-
-    return options;
-  }
-
-  async verifyRegistration(
-    email: string,
-    response: RegistrationResponseJSON,
-  ): Promise<LoginResultType> {
-    const user = await this.userRepository.findUserByEmail(email);
-    const expectedChallenge = await this.redisService.getOrThrow<string>(
-      user.uuid,
-      {
-        prefix: this.passkeyPrefix,
-      },
-    );
-
-    const { verified, registrationInfo } = await verifyRegistrationResponse({
-      response,
-      expectedChallenge,
-      expectedOrigin: this.passkeyRpOrigin,
-      expectedRPID: this.passkeyRpId,
-    });
-
-    if (!verified || !registrationInfo) {
-      throw new UnauthorizedException();
-    }
-
-    const { id, publicKey, counter } = registrationInfo.credential;
-
-    await this.authRepository.saveAuthenticator({
-      credentialId: id,
-      publicKey,
-      counter,
-      userUuid: user.uuid,
-    });
-
-    const refreshToken: string = this.generateOpaqueToken();
-    await this.redisService.set<Pick<User, 'uuid'>>(
-      refreshToken,
-      {
-        uuid: user.uuid,
-      },
-      {
-        prefix: this.refreshTokenPrefix,
-        ttl: this.refreshTokenExpireTime,
-      },
-    );
-    return {
-      accessToken: this.jwtService.sign({}, { subject: user.uuid }),
-      refreshToken,
-      accessTokenExpireTime: this.accessTokenExpireTime,
-      refreshTokenExpireTime: this.refreshTokenExpireTime,
-    };
-  }
-
   async authenticateOptions(
     email: string,
   ): Promise<PublicKeyCredentialRequestOptionsJSON> {
-    const user = await this.userRepository.findUserByEmail(email);
+    const user = await this.authRepository.findUserByEmail(email);
 
     if (!user || user.authenticators.length === 0) {
       throw new NotFoundException();
@@ -231,7 +149,7 @@ export class AuthService {
     email: string,
     response: AuthenticationResponseJSON,
   ): Promise<LoginResultType> {
-    const user = await this.userRepository.findUserByEmail(email);
+    const user = await this.authRepository.findUserByEmail(email);
     if (!user) throw new NotFoundException();
 
     const expectedChallenge = await this.redisService.getOrThrow<string>(
