+++ src/auth/types/userWithAuthenticators.ts
@@ -1,0 +1,7 @@
+import { Prisma } from '@prisma/client';
+
+export type UserWithAuthenticators = Prisma.UserGetPayload<{
+  include: {
+    authenticators: true;
+  };
+}>;
